% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkDate}
\alias{checkDate}
\title{Is the date within the data's 5 days range?}
\usage{
checkDate(date)
}
\arguments{
\item{date}{user specified date}
}
\value{
\code{checkDate} returns error
\code{"date should be either today or one of the next 4 days"}
if date is not either today's date or one of the next four days' date
}
\description{
This function checks if the specified date is within weather data 5 days
(i.e. today's date or one of the next 4 days)
since the weather forecast data is only available for these dates.
}
\examples{
# checkDate("2014-11-29")
}

