% Generated by roxygen2 (4.0.2): do not edit by hand
\name{goHike}
\alias{goHike}
\title{Is it a good day for hiking?}
\usage{
goHike(city = "Vancouver", date = Sys.Date())
}
\arguments{
\item{city}{The city you want to go Hiking in.
The default value is "Vancouver".}

\item{date}{The day you want to go Hiking on.
The default value is today's date. Please mind time zone difference.}
}
\value{
\code{goHike} returns

\code{"It's a lovely day for Hiking! :D"}
if chances of rain is less than 20 percent,

\code{"It might rain today! You might want to stay home"}
if chances of rain is between 20 percent and 60 percent, and

\code{"Ops! It's a rainy day, you better stay home :("} otherwise.
}
\description{
This function tells you if it's a good day for hiking based on Chances of Rain.
}
\details{
If you want to check if you can go hiking today or on the next four days
in any city around the world based on weather conditions (i.e chances of rain),
you don't have to go browse the Forecast news.
Instead, You can now ask R directly, without tempting yourself
by firing up your web browser.
}
\note{
case in \code{city} is ignored.

Make sure the entered \code{city} is spelled correctly.
As for cities that are composed of two or more words,
add underscore between each word (e.g. "Los_Angeles").

\code{date} should be today's or one of the next four days' date in "yyyy-mm-dd" format.
}
\examples{
# checks if you can go hiking today in Vancouver
goHike()

# checks if you can go hiking today in Seattle
goHike("Seattle")
}
\keyword{misc}

