% Generated by roxygen2 (4.0.2): do not edit by hand
\name{internet_connection}
\alias{internet_connection}
\title{Are you connected to the internet?}
\usage{
internet_connection()
}
\arguments{
\item{takes}{no argument}
}
\value{
\code{internet_connection} returns a logical that is only TRUE
if an internet connection can be established
}
\description{
This function checks if the user is connected to the internet.
}
\examples{
# internet_connection()
}

