% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{weatherR}
\alias{weatherR}
\alias{weatherR-package}
\title{Homework 10 R package: weatherR}
\description{
Allows you to access 5 days (i.e. today's and the next four days')
weather forecast information such as temperature in C, temperature in F,
Chance of rain, Humidity, Wind, etc in R.
}
\details{
This package accesses worldweatheronline.com to check weather forecast
(i.e temperature, chances of rain) for 5 days (today's or the next four days'),
and let you know if you can go hiking on that day based on chances of rain.
}

