% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wu_4day.R
\name{wu_4day}
\alias{wu_4day}
\title{weatherr}
\usage{
wu_4day()
}
\description{
This function when put inside a try catch function sends an email notification when there is an error
}
\examples{
wu_4day()
}
\keyword{catch,}
\keyword{email}
\keyword{error,}
\keyword{try}
