% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wu_all.R
\name{wu_all}
\alias{wu_all}
\title{weatherr}
\usage{
wu_all()
}
\description{
This function when put inside a try catch function sends an email notification when there is an error
}
\examples{
wu_all()
}
\keyword{catch,}
\keyword{email}
\keyword{error,}
\keyword{try}
