% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wu_current.R
\name{wu_current}
\alias{wu_current}
\title{weatherr}
\usage{
wu_current()
}
\description{
This function when put inside a try catch function sends an email notification when there is an error
}
\keyword{catch,}
\keyword{email}
\keyword{error,}
\keyword{try}
