% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getForecast}
\alias{getForecast}
\title{What is the forecast for today's weather?}
\usage{
getForecast(city = "Vancouver", date = Sys.Date())
}
\arguments{
\item{city}{The city you want to check its weather forecast.}

\item{date}{The day you want to check its weather forecast.}
}
\value{
\code{getForecast} returns \code{data.frame} that includes weather information
including Temperature in C, Weather Condition and Chances of Rain.
}
\description{
This function returns 5 days (today's and the next four days') weather information
including Temperature in C, Weather Condition and Chances of Rain.
}
\details{
If you want to check the weather forecast for today or the next 4 days
for any city around the world, you don't have to go browse the Forecast news.
Instead, You can now ask R directly, without tempting yourself
by firing up your web browser.
}
\note{
case in \code{city} is ignored

\code{date} should be today's or one of the next four days' date in "yyyy-mm-dd" formate
}
\examples{
# returns today's weather forecast for Vancouver
getForecast()

# returns today's weather forecast for Seattle
getForecast("Seattle")
}
\keyword{misc}

