\name{weatherr-package}
\alias{weatherr-package}
\alias{weatherr}
\docType{package}
\title{Tools for handling and scrapping instant weather feeds}
\description{Handling and scrapping instant weather forecasts and geographical information  }
\details{
\tabular{ll}{
Package: \tab weatherr\cr
Type: \tab Package\cr
Version: \tab 0.1.2\cr
Date: \tab 2015-09-08\cr
License: \tab GPL (>=2)\cr
}
}
\author{Stan Yip}